pci: accept MMIO pools whose exclusive end is 4 GiB
A pool ending exactly at 0x100000000 (e.g. 0xC0000000 + 0x40000000,
the classic top-half 32-bit MMIO layout) was a working configuration:
the old base + length - 1 initialization wrapped to 0xFFFFFFFF in
32-bit arithmetic. The overflow guard from the exclusive-limit fix
rejected such pools with base + length > 0xFFFFFFFF, aborting
enumeration - and the FSP caller discards the return value, so the
platform would boot with no PCI BARs programmed.

The limit fields cannot hold the exclusive end 0x100000000 while
32-bit, so widen mem_limit and mem_pf_limit (and the limit
parameters of pci_enum_next_aligned32 and pci_align_check_up, plus
the local in pci_program_bar) to 64-bit, and reject only pools whose
end is above the 32-bit space. The initialization now casts to
64-bit before the addition so the sum cannot wrap.

The T10xx PCIe setup initializes the same struct with the old
inclusive base + length - 1 form; align it to the exclusive
semantics the allocator enforces, or the last byte of the configured
pool is unusable.

New unit-pci-4gib build of the existing test file with the MMIO
pool [0xC0000000, 0x100000000): pci_enum_do() must accept the pool
and map a 1 MB BAR at the pool base. Fails on the old guard.

+++ include/pci.h
@@ -90,10 +90,12 @@ typedef  struct {
 
 struct pci_enum_info {
     uint32_t mem;
-    uint32_t mem_limit;
+    /* Exclusive pool ends.  64-bit: a pool may end exactly at 4 GiB
+     * (0x100000000), which a 32-bit field cannot represent. */
+    uint64_t mem_limit;
     uint32_t io;
     uint32_t mem_pf;
-    uint32_t mem_pf_limit;
+    uint64_t mem_pf_limit;
     uint8_t curr_bus_number;
 };
